moved grobid trainer utils to separate module
